backend/requirements.txt
# Backend requirements (Desktop PC)
paho-mqtt>=1.6,<2.0
opencv-python>=4.5
ultralytics>=8.0
transformers>=4.30
torch>=2.0
torchvision>=0.15
Pillow>=9.0
numpy>=1.24
